refactor(condo): DOMA-9619 moved function to helpers
