+++ apps/condo/domains/banking/schema/BankAccountReportTask.test.js
@@ -32,6 +32,7 @@ const {
     createTestBankCategory,
     createTestBankCostItem,
 } = require('@condo/domains/banking/utils/testSchema')
+const { makeBankAccountWithData } = require('@condo/domains/banking/utils/testSchema/bankAccount')
 const { HOLDING_TYPE } = require('@condo/domains/organization/constants/common')
 const {
     createTestOrganization,
@@ -44,69 +45,6 @@ const {
     makeClientWithSupportUser,
 } = require('@condo/domains/user/utils/testSchema')
 
-
-const prepareBankAccountData = async (client, o10n, bankIntegration, category) => {
-    const currentDate = dayjs()
-    const formattedCurrentDate = currentDate.format('YYYY-MM-DD')
-
-    const incomeTransactions = 5
-    const incomeTransactionAmount = 50
-    const totalIncome = incomeTransactions * incomeTransactionAmount
-
-    const outcomeTransactions = 5
-    const outcomeTransactionAmount = 10
-    const totalOutcome = outcomeTransactions * outcomeTransactionAmount
-
-    const totalAmount = totalIncome - totalOutcome
-
-    const [integrationContext] = await createTestBankIntegrationAccountContext(client, bankIntegration, o10n, {
-        meta: {
-            amount: String(totalAmount),
-            amountAt: formattedCurrentDate,
-        },
-    })
-    const [account] = await createTestBankAccount(client, o10n, {
-        integrationContext: { connect: { id: integrationContext.id } },
-    })
-
-    const [costItem] = await createTestBankCostItem(client, category, {
-        isOutcome: false,
-    })
-
-    const [contractorAccount] = await createTestBankContractorAccount(client, o10n, {
-        costItem: { connect: { id: costItem.id } },
-    })
-
-    for (let i = 0; i < incomeTransactions; i++) {
-        await createTestBankTransaction(client, account, contractorAccount, integrationContext, o10n, {
-            date: formattedCurrentDate,
-            amount: String(incomeTransactionAmount),
-            isOutcome: false,
-        })
-
-    }
-    for (let i = 0; i < outcomeTransactions; i++) {
-        await createTestBankTransaction(client, account, { id: faker.datatype.uuid() }, integrationContext, o10n, {
-            date: formattedCurrentDate,
-            amount: String(outcomeTransactionAmount),
-            isOutcome: true,
-            contractorAccount: undefined,
-        })
-    }
-
-    return {
-        integrationContext,
-        account,
-        costItem,
-        contractorAccount,
-        bankIntegration,
-        category,
-        formattedCurrentDate,
-        totalIncome,
-        totalOutcome,
-        totalAmount,
-    }
-}
 
 describe('BankAccountReportTask', () => {
     let admin, bankIntegration, category
@@ -817,9 +755,9 @@ describe('BankAccountReportTask', () => {
         it('Should generate correct reports for different bank accounts', async () => {
             const [org] = await createTestOrganization(admin)
 
-            const bankAccountData1 = await prepareBankAccountData(admin, org, bankIntegration, category)
-            const bankAccountData2 = await prepareBankAccountData(admin, org, bankIntegration, category)
-            const bankAccountData3 = await prepareBankAccountData(admin, org, bankIntegration, category)
+            const bankAccountData1 = await makeBankAccountWithData(admin, org, bankIntegration, category)
+            const bankAccountData2 = await makeBankAccountWithData(admin, org, bankIntegration, category)
+            const bankAccountData3 = await makeBankAccountWithData(admin, org, bankIntegration, category)
             const bankAccountsData = [bankAccountData1, bankAccountData2, bankAccountData3]
 
             const [createdReportTask1] = await createTestBankAccountReportTask(admin, bankAccountData1.account, org, admin.user.id, { progress: 0 })
